prac_vectorstore/pyproject.toml
[tool.poetry]
name = "prac-vectorstore"
version = "0.1.0"
description = ""
authors = ["JeongJaeSoon <wjdwotns1006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.285"
python-dotenv = "^1.0.0"
chromadb = "^0.4.9"
tiktoken = "^0.4.0"
wikipedia = "^1.4.0"
openai = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
